requirements.txt
annotated-types>=0.7.0
anyio>=4.9.0
appnope
asttokens
cachetools>=5.5.2
certifi>=2025.1.31
charset-normalizer>=3.4.1
click>=8.1.8
comm
debugpy
decorator
dotenv>=0.9.9
exceptiongroup
executing
fastapi>=0.115.12
google-ai-generativelanguage>=0.6.15
google-api-core>=2.24.2
google-api-python-client>=2.167.0
google-auth>=2.39.0
google-auth-httplib2>=0.2.0
google-generativeai>=0.8.5
googleapis-common-protos>=1.70.0
grpcio>=1.72.0rc1
grpcio-status>=1.71.0
h11>=0.14.0
httplib2>=0.22.0
idna>=3.10
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jedi
jupyter_client
jupyter_core
matplotlib-inline
nest_asyncio
numpy>=1.26.4
opencv-python>=4.11.0.86
packaging
pandas>=2.2.3
parso
pdf2image>=1.17.0
pexpect
pickleshare
pillow>=11.1.0
platformdirs
prompt_toolkit
proto-plus>=1.26.1
protobuf>=5.29.4
psutil
ptyprocess
pure_eval
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pydantic>=2.11.3
pydantic_core>=2.33.1
Pygments
pyparsing>=3.2.3
pytesseract>=0.3.13
python-dateutil
python-dotenv>=1.1.0
python-multipart>=0.0.20
pytz>=2025.1
pyzmq
requests>=2.32.3
rsa>=4.9.1
six
sniffio>=1.3.1
stack_data
starlette>=0.46.2
tornado
tqdm>=4.67.1
traitlets
typing-inspection>=0.4.0
typing_extensions
tzdata>=2025.1
uritemplate>=4.1.1
urllib3>=2.4.0
uvicorn>=0.34.2
wcwidth
zipp